Location Base Application
Tracking system is used for the observing of persons or objects location in motion or stands still and supplies periodically sequence of location data to a monitoring system. Nowadays, tracking systems are now playing a key role in domestic and industrial area. There is range of tracking systems available but no single low power system that integrates all tracking of any movable objects. The present work is to develop an Android based Smart tracking system where authorized user gets all status of device on his mobile phone. In this project, Advance Encryption Standard algorithm (AES) is implemented to encrypt GPS and sensor parameters. DragonBoard410c is used as SBC which has a high performance Qualcomm Snapdragon processor used for IoT platform. The presented work concentrates on the wireless Encrypted GPS and sensor data communication between the device and user’s mobile via Internet and how to access and monitor sensor parameters. This tracking system is low in size, power consumption.